refactor: move login to _auth layout
